Update .gitignore to conform to accepted standards
It is standard practice for the repo-wide .gitignore to feature files that are relevant to the repo itself, not to the various editors, etc. that an individual programmer may use to work on the repo. As such, all temporary files generated by text editors have been removed from the repo's .gitingnore, and our team will be expected to use their own global .gitignore to manage editor-specific files.